+++ python/fnllm/tests/unit/llm/services/test_retryer.py
@@ -29,7 +29,7 @@ async def test_retrying_llm_passthrough():
     retryer = TestRetryer(
         retryable_errors=[],
         events=events,
-        retry_strategy=RetryStrategy.TENACITY,
+        retry_strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
         retryable_error_handler=None,
     )
     llm = retryer.decorate(delegate)
@@ -75,7 +75,7 @@ async def test_retrying_llm_recovers():
     retryer = TestRetryer(
         retryable_errors=[ValueError],
         events=events,
-        retry_strategy=RetryStrategy.TENACITY,
+        retry_strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
         retryable_error_handler=None,
     )
     llm = retryer.decorate(delegate)
@@ -104,7 +104,7 @@ async def test_retrying_llm_raises_retries_exhausted():
         max_retry_wait=0.1,
         max_retries=5,
         events=LLMEvents(),
-        retry_strategy=RetryStrategy.TENACITY,
+        retry_strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
         retryable_error_handler=None,
     )
     llm = retryer.decorate(delegate)
@@ -124,7 +124,7 @@ async def test_retrying_llm_emits_error_if_not_retryable():
         max_retry_wait=0.1,
         max_retries=5,
         events=LLMEvents(),
-        retry_strategy=RetryStrategy.TENACITY,
+        retry_strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
         retryable_error_handler=None,
     )
     llm = retryer.decorate(delegate)