+++ python/fnllm/CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add new RetryStrategy enum values:
+	* `RetryStrategy.EXPONENTIAL_BACKOFF` (default strategy) - for exponential backoff wait times 
+	* `RetryStrategy.RANDOM_WAIT` - for random wait times between `[0, config.max_wait]`.
+	* `RetryStrategy.INCREMENTAL_WAIT` - for linear incremental wait times between `[0, config.max_wait]`.
 ### Changed
 ### Deprecated
+- Deprecate `RetryStrategy.TENACITY` in favor of `RetryStrategy.EXPONENTIAL_BACKOFF`.
 ### Removed
 ### Fixed
 * Fix issue with TPM Limiter where requests with tokens > tokens_per_minute limit would automatically fail.